riveredge-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riveredge-core"
version = "1.0.0"
description = "RiverEdge SaaS 多租户框架 - 后端核心系统"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",  # redis-p, 使用 redis.asyncio 异步接口
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[tool.aerich]
tortoise_orm = "core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"
